requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
openai==1.12.0
google-genai>=1.0.0
google-auth>=2.47.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.118.0
gspread==6.0.2
python-dotenv==1.0.1
pydantic>=2.9.0
jinja2==3.1.3
aiofiles==23.2.1
python-multipart==0.0.9
httpx>=0.28.1
